+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 - Add argument (export_csv) to the build function to write the dataset as a CSV file
 - Add argument (output_path) a directory path for building datasets in
 - Add DataType: random_float to generate random floats
+- Add Argument: (schema_dict) to read a schema from a dictionary
 
 ### Changed
 - Fixed a bug in data type: Decrement, where it was adding extra row